fix(ffi): redefine win_col_off behavior to replace removed function, fixes #38
As noted by @cassava, as of a recent version of this plugin, the
`curwin_col_off` function was removed, leaving us to fend for ourselves.
Luckily, @cassava also proposed a fix, the likes of which seems to work on my
machine and from the limited reports, other individuals' machines as well.

This commit simply implements the patch specified by @cassava as the solution to
issues related to `curwin_col_off` no longer existing.